move scheduling code from dirty_scope to scheduler
